control wave frequence
we want to control the frequency on x and z
but we want to control them separately

In the vertex shader, retrieve the uniform —be careful, it's a vec2— and apply it in the sin(...) with only the x property to begin with:

// ...
uniform vec2 uBigWavesFrequency;

void main()
{
    // ...

    float elevation = sin(modelPosition.x * uBigWavesFrequency.x) * uBigWavesElevation;

    // ...
}

This should result in more waves because the frequency is higher.

Let's use the second value of the uBigWavesFrequency (y) to control the waves on the z axis. We can do that by multiplying the first sin(...) by another sin(...):

float elevation = sin(modelPosition.x * uBigWavesFrequency.x) * sin(modelPosition.z * uBigWavesFrequency.y) * uBigWavesElevation;

You should get waves on the z axis too.