Move the shaders to separate files
Now is an excellent time to move the shaders to separate files before they get too long and unmanageable.

In /src/, create a shaders/, and then a galaxy/ folder inside.

Inside that folder, create a vertex.glsl file with the vertexShader property content:

The result should be the same.

You don't need to put the shaders in separate files, but it's good practice, and the syntax coloration might prevent you from making mistakes.

Also, if you installed a linter like suggested in a previous lesson, you'll see potential errors before refreshing.