random size and calibrate for pixel ratio
In real life, stars have different sizes. Let's add some randomness. We want to associate a distinct value for each vertex. We are going to use an attribute.

Add a aScale attribute to the geometry. We already have a position, and a color attribute and we can easily add the new attribute following the same instructions:

Make sure to use 1 instead of 3 when creating the Float32Array and the BufferAttribute because this value is a float and not a vec3 like the others —we need only one value per vertex. We also named the attribute aScale with a prepended a.

You might be tempted to change the position and color attributes for aPosition and aColor, but that would result in a bug because we are using a ShaderMaterial which pre-pend code to our vertex shader and that code add something like attribute vec3 position; and attribute vec3 color;.

We can now retrieve the attribute in the vertex shader and multiply the uSize by it:

You should see particles with random sizes.

Fixing the pixel ratio
However, we have a problem with our particles. Their size depends on the screen's pixel ratio. Remember that we updated the pixel ratio of the renderer by using the following line:

renderer.setPixelRatio(Math.min(window.devicePixelRatio, 2))
JavaScript
If you have a screen with a pixel ratio of 1, the particles will look 2 times bigger than if you had a screen with a pixel ratio of 2.

We need a solution to get the same particle size regardless of the pixel ratio.

They are multiple ways of doing so. The easiest one is to multiply the uSize value by the pixel ratio of the renderer. We can retrieve this pixel ratio with the getPixelRatio() method:

method:

material = new THREE.ShaderMaterial({

    // ...

    uniforms:
    {
        uSize: { value: 8 * renderer.getPixelRatio() }
    }

    // ...
})
JavaScript
Unhappily, this code won't work because we are creating the material before we create the renderer. To fix that, simply move the first call of generateGalaxy after the instantiating the renderer:

/**
 * Renderer
 */
const renderer = new THREE.WebGLRenderer({
    canvas: canvas
})

// ...

/**
 * Generate the first galaxy
 */
generateGalaxy()
JavaScript